% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR functions.R
\name{IntraTime}
\alias{IntraTime}
\title{Calculate Time Period Represented by the Intradermal Section of the Whisker}
\usage{
IntraTime(L, L.int, L.asym, k)
}
\arguments{
\item{L}{numeric; length of the entire whisker minus the intradermal section}

\item{L.int}{numeric; intradermal length of the whisker}

\item{L.asym}{numeric; asymptotic length}

\item{k}{numeric; growth coefficient, calculate using the function k}
}
\value{
The value returned is the number of days represented by the intradermal section of the whisker
}
\description{
\code{IntraTime} is used to calculate the time period represented by the intradermal section of the whisker
}
\examples{
IntraTime(L = 140, L.int = 10, L.asym = 150, k = 0.0126)
}
\author{
The function was written by Gary Truong with collaboration from Ben Walker
and Anna Lewis from the University of New South Wales
}
\keyword{intradermal}
\keyword{time}
\keyword{von-Bertanlanffy}
