% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR functions.R
\name{LP1}
\alias{LP1}
\title{Calculate Segment Length}
\usage{
LP1(L, L.asym, Time, k)
}
\arguments{
\item{L}{Length of the cut whisker}

\item{L.asym}{Asymptotic length}

\item{Time}{Time period to be represented by the cut segment in days}

\item{k}{Growth coefficient, calculate using the function k}
}
\value{
Where to cut along the whisker
}
\description{
This function allows you calculate where to cut to represent a specific time period.
}
\examples{
LP1(L=100, L.asym=150,Time=28 k=0.0126)

}
\keyword{length}
\keyword{segment}
\keyword{whisker}
