% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR functions.R
\name{MaxTime}
\alias{MaxTime}
\title{Calculate Maximum Time represented by sample}
\usage{
MaxTime(L, L.asym, k)
}
\description{
This function uses the simple Von Bertanlanffy equations to calculate the time represented by the entire sample
}
