% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR_functions.R
\name{section}
\alias{section}
\title{Generate a Table of Section Positions}
\usage{
section(L, L.asym, k, Time = 1)
}
\arguments{
\item{L}{numeric; the starting length of the sample}

\item{k}{numeric; growth coefficient, calculate using the function k}

\item{Time}{numeric; period of growth to be represented by each section. Default is 1 day.}

\item{L.asym.}{numeric; asymptotic length}
}
\description{
\code{section} uses the LP1 function to generate a tibble of all the sections legnths representing 1 day of growth.
}
