% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR functions.R
\name{TP1}
\alias{TP1}
\title{Calculate Time Represented by Segment}
\usage{
TP1(L, L.asym, LP1, k)
}
\arguments{
\item{L}{Length of the cut whisker}

\item{L.asym}{Asymptotic length}

\item{LP1}{Length of the whisker minus the cut sampled segment}

\item{k}{Growth coefficient, calculate using the function k}
}
\value{
Time represented by the sampled segment in days
}
\description{
This function allows you calculate how much time is represented by a sampled section of whisker.
It uses the discrete step by step Von Bertanlanffy equation
}
\examples{
TP1(L=100, L.asym=150, LP1=90, k=0.0126)

}
\keyword{segment}
\keyword{time}
\keyword{whisker}
