% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR functions.R
\name{k}
\alias{k}
\title{Calculate k growth coefficient}
\usage{
k(ML, L.asym, Time)
}
\arguments{
\item{ML}{Maximum length}

\item{L.asym}{Asymptotic length}

\item{Time}{Estimated time to reach maximum length}
}
\value{
Time represented by the sampled segment in days
}
\description{
Calculate k growth coefficient
}
\examples{
k(L= 100, )

}
\keyword{coefficient}
\keyword{growth}
\keyword{k}
