% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiskR_functions.R
\name{maxTime}
\alias{maxTime}
\title{Calculate Maximum Time Represented by Sample}
\usage{
maxTime(L, L.asym, k)
}
\arguments{
\item{L}{numeric; length of the entire sample}

\item{L.asym}{numeric; asymptotic length}

\item{k}{numeric; growth coefficient, calculate using the function k}
}
\value{
The value returned is the number of days represented by the entire sample
}
\description{
\code{maxTime} uses the simple Von Bertanlanffy equations to calculate the time represented by the entire sample
}
\examples{
maxTime(L = 140, L.asym = 150, k = 0.0126)
}
\author{
The function was written by Gary Truong with collaboration from Ben Walker
and Anna Lewis from the University of New South Wales
}
\keyword{intradermal}
\keyword{time}
\keyword{von-Bertanlanffy}
